refactor : getById 파라미터 수정(int -> long)

+++ src/main/java/roomescape/dao/ReservationDAO.java
@@ -11,6 +11,6 @@ public interface ReservationDAO {
 
     void delete(long id);
 
-    Reservation getById(int id);
+    Reservation getById(long id);
 
 }
